refactor(PlayRecord): RecordFactory 역할 PlayRecord에 통합
