+++ src/main/java/blackjack/domain/PlayRecord.java
@@ -1,6 +1,13 @@
 package blackjack.domain;
 
 import static blackjack.domain.PlayStatus.*;
+import static java.util.stream.Collectors.*;
+
+import java.util.List;
+import java.util.Map;
+
+import blackjack.domain.participant.Name;
+import blackjack.domain.participant.Player;
 
 public enum PlayRecord {
 
@@ -9,7 +16,13 @@ public enum PlayRecord {
     LOSS,
     BLACKJACK;
 
-    public static PlayRecord of(int dealerScore, int score, boolean isBlackjack) {
+    public static Map<Name, PlayRecord> createPlayRecords(List<Player> players, int dealerScore) {
+        return List.copyOf(players).stream()
+            .collect(toUnmodifiableMap(Player::getName,
+                player -> of(dealerScore, player.getScore(), player.isBlackjack())));
+    }
+
+    private static PlayRecord of(int dealerScore, int score, boolean isBlackjack) {
         if (isPlayerLoss(dealerScore, score)) {
             return LOSS;
         }
